Fixes to build scripts.

+++ install-scripts/mac/fink-daily-build-packages.sh
@@ -75,8 +75,10 @@ tar -xzf $FINK_ROOT/fink/dists/biolab/main/finkinfo/all.tgz -C $FINK_ROOT/fink/d
 rm -f $FINK_ROOT/fink/dists/biolab/main/finkinfo/all.tgz
 
 # Move info files from local/main/finkinfo (put there by dailyru[-finkonly].sh
-echo "Updating new fink info files."
-mv $FINK_ROOT/fink/dists/local/main/finkinfo/*.info $FINK_ROOT/fink/dists/biolab/main/finkinfo/
+if [ -e $FINK_ROOT/fink/dists/local/main/finkinfo/*.info ]; then
+	echo "Updating new fink info files."
+	mv $FINK_ROOT/fink/dists/local/main/finkinfo/*.info $FINK_ROOT/fink/dists/biolab/main/finkinfo/
+fi
 
 if ! grep '^Trees:' $FINK_ROOT/etc/fink.conf | grep -q 'biolab/main'; then
 	echo "Adding local biolab Fink info files repository to Fink configuration."
